verifyvehicle
"If one is looking for a platform where all the information about the online vehicle verification and registration of every type of vehicle in Pakistan is available, then the Motor Transport Management Information System (MTMIS) is designed for these citizens. It is the latest digital platform where the registration status of all types of vehicles in Pakistan is available. For more information, visit verifyvehicle.pk."